chore: add tsdoc comments
