+++ src/sdk/modules/ownables/decorators/getRemoveOwnerTx.ts
@@ -1,20 +1,41 @@
 import { getRemoveOwnableValidatorOwnerAction } from "@rhinestone/module-sdk"
 import type { Chain, Client, Hex, PublicClient, Transport } from "viem"
-import type { SmartAccount } from "viem/account-abstraction"
+import type { ModularSmartAccount } from "../../utils/Types"
+
 import { parseAccount } from "viem/utils"
 import { AccountNotFoundError } from "../../../account/utils/AccountNotFound"
 import type { Call } from "../../../account/utils/Types"
 
-export type GetRemoveOwnerTxParameters<TSmartAccount> = {
-  account?: TSmartAccount
+/**
+ * Parameters for getting the transaction to remove an owner from a smart account.
+ *
+ * @template TModularSmartAccount - The type of the smart account, which can be a ModularSmartAccount or undefined.
+ */
+export type GetRemoveOwnerTxParameters<TModularSmartAccount> = {
+  /** The smart account to remove the owner from. If not provided, the client's account will be used. */
+  account?: TModularSmartAccount
+  /** The address of the owner to be removed. */
   owner: Hex
 }
 
+/**
+ * Generates the transaction data for removing an owner from a smart account.
+ *
+ * This function prepares the necessary transaction data to remove an owner from the specified smart account.
+ * It doesn't send the transaction, but returns the data needed to do so.
+ *
+ * @template TModularSmartAccount - The type of the smart account, which can be a ModularSmartAccount or undefined.
+ * @param client - The client used to interact with the blockchain.
+ * @param parameters - The parameters for removing the owner.
+ * @returns A promise that resolves to a Call object containing the transaction data.
+ * @throws {AccountNotFoundError} If no account is provided and the client doesn't have an associated account.
+ * @throws {Error} If there's an error getting the remove owner action or if the public client is not found.
+ */
 export async function getRemoveOwnerTx<
-  TSmartAccount extends SmartAccount | undefined
+  TModularSmartAccount extends ModularSmartAccount | undefined
 >(
-  client: Client<Transport, Chain | undefined, TSmartAccount>,
-  parameters: GetRemoveOwnerTxParameters<TSmartAccount>
+  client: Client<Transport, Chain | undefined, TModularSmartAccount>,
+  parameters: GetRemoveOwnerTxParameters<TModularSmartAccount>
 ): Promise<Call> {
   const { account: account_ = client.account, owner } = parameters
 
@@ -24,7 +45,7 @@ export async function getRemoveOwnerTx<
     })
   }
 
-  const account = parseAccount(account_) as SmartAccount
+  const account = parseAccount(account_) as ModularSmartAccount
   const publicClient = account.client
 
   if (!publicClient) {
