+++ src/sdk/modules/ownables/decorators/prepareSignatures.ts
@@ -1,20 +1,20 @@
 import type { Chain, Client, Hex, PublicClient, Transport } from "viem"
-import type { SmartAccount } from "viem/account-abstraction"
 import { encodePacked, parseAccount } from "viem/utils"
 import { AccountNotFoundError } from "../../../account/utils/AccountNotFound"
+import type { ModularSmartAccount } from "../../utils/Types"
 
 export type PrepareSignaturesParameters<
-  TSmartAccount extends SmartAccount | undefined
+  TModularSmartAccount extends ModularSmartAccount | undefined
 > = {
-  account?: TSmartAccount
+  account?: TModularSmartAccount
   signatures: Hex[]
 }
 
 export async function prepareSignatures<
-  TSmartAccount extends SmartAccount | undefined
+  TModularSmartAccount extends ModularSmartAccount | undefined
 >(
-  client: Client<Transport, Chain | undefined, TSmartAccount>,
-  parameters: PrepareSignaturesParameters<TSmartAccount>
+  client: Client<Transport, Chain | undefined, TModularSmartAccount>,
+  parameters: PrepareSignaturesParameters<TModularSmartAccount>
 ): Promise<Hex> {
   const { account: account_ = client.account, signatures } = parameters
 
@@ -24,7 +24,7 @@ export async function prepareSignatures<
     })
   }
 
-  const account = parseAccount(account_) as SmartAccount
+  const account = parseAccount(account_) as ModularSmartAccount
   const publicClient = account?.client as PublicClient
 
   if (!publicClient) {
