+++ src/sdk/modules/ownables/toOwnables.ts
@@ -13,23 +13,47 @@ import {
   decodeAbiParameters,
   encodeAbiParameters
 } from "viem"
-import type { SmartAccount } from "viem/account-abstraction"
-import type { Module, ModuleParameters } from "../utils/Types"
+import type {
+  ModularSmartAccount,
+  Module,
+  ModuleParameters
+} from "../utils/Types"
 import { type ToModuleParameters, toModule } from "../utils/toModule"
 
+/**
+ * Parameters for creating an Ownable module.
+ * Extends ToModuleParameters but replaces 'accountAddress' with 'account'.
+ */
 type ToOwnableModuleParameters = Omit<ToModuleParameters, "accountAddress"> & {
-  account: SmartAccount
+  /** The modular smart account to associate with this module. */
+  account: ModularSmartAccount
+  /** Optional initialization arguments for the module. */
   moduleInitArgs?: GetOwnablesModuleInitDataParams
+  /** Optional public client for blockchain interactions. */
   client?: PublicClient
 }
 
+/**
+ * Parameters for initializing the Ownables module.
+ */
 export type GetOwnablesModuleInitDataParams = {
+  /** The threshold number of signatures required for operations. */
   threshold: bigint
+  /** Array of owner addresses for the module. */
   owners: Address[]
 }
 
+/**
+ * Parameters specific to the Ownable module.
+ */
 export type OwnableModuleParameters = ModuleParameters
 
+/**
+ * Generates the initialization data for the Ownables module.
+ *
+ * @param parameters - The parameters for initializing the module.
+ * @returns The module metadata including the address, type, and encoded init data.
+ */
 export const getOwnablesModuleInitData = (
   parameters: GetOwnablesModuleInitDataParams
 ): ModuleMeta => ({
@@ -43,9 +67,43 @@ export const getOwnablesModuleInitData = (
     [parameters.threshold, parameters.owners]
   )
 })
+
+/**
+ * Generates the initialization data for the Ownables module.
+ * This function currently returns an empty hex string.
+ *
+ * @param _ - Optional initialization parameters (currently unused).
+ * @returns An empty hex string.
+ */
 export const getOwnablesInitData = (_?: GetOwnablesModuleInitDataParams): Hex =>
   "0x"
 
+/**
+ * Creates an Ownable module for a modular smart account.
+ *
+ * This function sets up an Ownable module with the specified parameters,
+ * including threshold and owners for the smart account.
+ *
+ * @param parameters - The parameters for creating the Ownable module.
+ * @returns A Module object representing the created Ownable module.
+ *
+ * @example
+ * ```typescript
+ * const ownableModule = toOwnables({
+ *   account: mySmartAccount,
+ *   signer: mySigner,
+ *   moduleInitArgs: {
+ *     threshold: 2n,
+ *     owners: ['0x123...', '0x456...']
+ *   }
+ * });
+ * ```
+ *
+ * @remarks
+ * - If the module is already installed, it will use the existing threshold.
+ * - If not installed, it will use the threshold from the initialization parameters.
+ * - The function generates a mock signature based on the threshold.
+ */
 export const toOwnables = (parameters: ToOwnableModuleParameters): Module => {
   const {
     account,
