+++ src/sdk/modules/smartSessions/Helpers.ts
@@ -18,6 +18,7 @@ import type {
   ActionConfig,
   RawActionConfig,
   Rule,
+  SessionData,
   SpendingLimitsParams
 } from "./Types"
 
@@ -25,6 +26,11 @@ const TIMEFRAME_POLICY_ADDRESS = addresses.TimeframePolicy
 
 export const MAX_RULES = 16
 
+/**
+ * Generates a random salt as a hexadecimal string.
+ *
+ * @returns A 32-byte hexadecimal string prefixed with '0x'.
+ */
 export const generateSalt = (): Hex => {
   const randomBytes = new Uint8Array(32)
   crypto.getRandomValues(randomBytes)
@@ -33,6 +39,13 @@ export const generateSalt = (): Hex => {
   ).join("")}` as Hex
 }
 
+/**
+ * Creates an ActionConfig object from rules and a value limit.
+ *
+ * @param rules - An array of Rule objects.
+ * @param valueLimit - The maximum value allowed for the action.
+ * @returns An ActionConfig object.
+ */
 export const createActionConfig = (
   rules: Rule[],
   valueLimit: bigint
@@ -44,6 +57,14 @@ export const createActionConfig = (
   valueLimitPerUse: valueLimit
 })
 
+/**
+ * Creates an ActionData object.
+ *
+ * @param contractAddress - The address of the contract.
+ * @param functionSelector - The function selector or AbiFunction.
+ * @param policies - An array of PolicyData objects.
+ * @returns An ActionData object.
+ */
 export const createActionData = (
   contractAddress: Address,
   functionSelector: string | AbiFunction,
@@ -58,6 +79,12 @@ export const createActionData = (
   }
 }
 
+/**
+ * Converts an ActionConfig to a RawActionConfig.
+ *
+ * @param config - The ActionConfig to convert.
+ * @returns A RawActionConfig object.
+ */
 export const toActionConfig = (config: ActionConfig): RawActionConfig => {
   // Ensure we always have 16 rules, filling with default values if necessary
   const filledRules = [...config.paramRules.rules]
@@ -94,6 +121,13 @@ export const toActionConfig = (config: ActionConfig): RawActionConfig => {
   }
 }
 
+/**
+ * Gets the permission ID for a given session.
+ *
+ * @param client - The PublicClient to use for the contract call.
+ * @param session - The Session object.
+ * @returns A promise that resolves to the permission ID as a Hex string.
+ */
 export const getPermissionId = async ({
   client,
   session
@@ -109,6 +143,14 @@ export const getPermissionId = async ({
   })) as Hex
 }
 
+/**
+ * Checks if a session is enabled for a given account.
+ *
+ * @param client - The PublicClient to use for the contract call.
+ * @param accountAddress - The address of the account.
+ * @param permissionId - The permission ID to check.
+ * @returns A promise that resolves to a boolean indicating if the session is enabled.
+ */
 export const isSessionEnabled = ({
   client,
   accountAddress,
@@ -125,6 +167,12 @@ export const isSessionEnabled = ({
     args: [permissionId, accountAddress]
   })
 
+/**
+ * Converts an ActionConfig to a UniversalActionPolicy.
+ *
+ * @param actionConfig - The ActionConfig to convert.
+ * @returns A PolicyData object representing the UniversalActionPolicy.
+ */
 export const toUniversalActionPolicy = (
   actionConfig: ActionConfig
 ): PolicyData => ({
@@ -134,6 +182,13 @@ export const toUniversalActionPolicy = (
   ])
 })
 
+/**
+ * Creates a TimeRangePolicy.
+ *
+ * @param validUntil - The timestamp until which the policy is valid.
+ * @param validAfter - The timestamp after which the policy is valid.
+ * @returns A PolicyData object representing the TimeRangePolicy.
+ */
 export const toTimeRangePolicy = (
   validUntil: number,
   validAfter: number
@@ -158,11 +213,20 @@ export const toTimeRangePolicy = (
   return timeFramePolicyData
 }
 
+/**
+ * A PolicyData object representing a sudo policy.
+ */
 export const sudoPolicy: PolicyData = {
   policy: "0x529Ad04F4D83aAb25144a90267D4a1443B84f5A6",
   initData: "0x"
 }
 
+/**
+ * Converts SpendingLimitsParams to a SpendingLimitsPolicy.
+ *
+ * @param params - An array of SpendingLimitsParams.
+ * @returns A PolicyData object representing the SpendingLimitsPolicy.
+ */
 export const toSpendingLimitsPolicy = (
   params: SpendingLimitsParams
 ): PolicyData => {
@@ -175,6 +239,9 @@ export const toSpendingLimitsPolicy = (
   }
 }
 
+/**
+ * An object containing policy conversion functions.
+ */
 export const policies = {
   to: {
     universalAction: toUniversalActionPolicy,
@@ -182,5 +249,25 @@ export const policies = {
   },
   sudo: sudoPolicy
 } as const
+
+/**
+ * Zips SessionData into a compact string representation.
+ *
+ * @param sessionData - The SessionData object to be zipped.
+ * @returns A string representing the zipped SessionData.
+ */
+export function zipSessionData(sessionData: SessionData): string {
+  return JSON.stringify(sessionData)
+}
+
+/**
+ * Unzips a string representation back into a SessionData object.
+ *
+ * @param zippedData - The string representing the zipped SessionData.
+ * @returns The unzipped SessionData object.
+ */
+export function unzipSessionData(zippedData: string): SessionData {
+  return JSON.parse(zippedData) as SessionData
+}
 
 export default policies
