+++ src/sdk/modules/smartSessions/decorators/useSession.ts
@@ -1,27 +1,68 @@
 import type { Chain, Client, Hex, Transport } from "viem"
-import { type SmartAccount, sendUserOperation } from "viem/account-abstraction"
+import { sendUserOperation } from "viem/account-abstraction"
 import { getAction, parseAccount } from "viem/utils"
 import type { NexusAccount } from "../../../account/toNexusAccount"
 import { AccountNotFoundError } from "../../../account/utils/AccountNotFound"
-import type { Execution } from "../../utils/Types"
+import type { Execution, ModularSmartAccount } from "../../utils/Types"
 
-// If the session is enabled for multiple actions, it is possible to send a batch transaction. hence it accepts an array of executions.
-// permisisonId corresponds to already enabled session.
+/**
+ * Parameters for using a smart session to execute actions.
+ *
+ * @template TModularSmartAccount - Type of the modular smart account, extending ModularSmartAccount or undefined.
+ */
 export type UseSessionParameters<
-  TSmartAccount extends SmartAccount | undefined
+  TModularSmartAccount extends ModularSmartAccount | undefined
 > = {
+  /** Array of executions to perform in the session. Allows for batch transactions if the session is enabled for multiple actions. */
   actions: Execution[]
+  /** The maximum fee per gas unit the transaction is willing to pay. */
   maxFeePerGas?: bigint
+  /** The maximum priority fee per gas unit the transaction is willing to pay. */
   maxPriorityFeePerGas?: bigint
+  /** The nonce of the transaction. If not provided, it will be determined automatically. */
   nonce?: bigint
-  account?: TSmartAccount
+  /** The modular smart account to use for the session. If not provided, the client's account will be used. */
+  account?: TModularSmartAccount
 }
 
+/**
+ * Executes actions using a smart session.
+ *
+ * This function allows for the execution of one or more actions within an enabled smart session.
+ * It can handle batch transactions if the session is configured for multiple actions.
+ *
+ * @template TModularSmartAccount - Type of the modular smart account, extending ModularSmartAccount or undefined.
+ * @param client - The client used to interact with the blockchain.
+ * @param parameters - Parameters for using the session, including actions to execute and optional gas settings.
+ * @returns A promise that resolves to the hash of the sent user operation.
+ *
+ * @throws {AccountNotFoundError} If no account is provided and the client doesn't have an associated account.
+ *
+ * @example
+ * ```typescript
+ * const result = await useSession(nexusClient, {
+ *   actions: [
+ *     {
+ *       target: '0x1234...',
+ *       value: 0n,
+ *       callData: '0xabcdef...'
+ *     }
+ *   ],
+ *   maxFeePerGas: 1000000000n
+ * });
+ * console.log(`Transaction hash: ${result}`);
+ * ```
+ *
+ * @remarks
+ * - Ensure that the session is enabled and has the necessary permissions for the actions being executed.
+ * - For batch transactions, all actions must be permitted within the same session.
+ * - The function uses the `sendUserOperation` method, which is specific to account abstraction implementations.
+ */
 export async function useSession<
-  TSmartAccount extends SmartAccount | undefined
+  TModularSmartAccount extends ModularSmartAccount | undefined
 >(
-  client: Client<Transport, Chain | undefined, TSmartAccount>,
-  parameters: UseSessionParameters<TSmartAccount>
+  client: Client<Transport, Chain | undefined, TModularSmartAccount>,
+  parameters: UseSessionParameters<TModularSmartAccount>
 ): Promise<Hex> {
   const {
     account: account_ = client.account,
